chore: UserActionを削除
